requirements.txt
flask==2.2.2
Werkzeug==2.2.2
spotipy==2.25.1
python-dotenv==0.21.1
openai==1.10.0